add log parser function

+++ LogParser/Models/Completion/CompletionRequest.cs
@@ -20,10 +20,10 @@ public class CompletionRequest : Request
     public int? MaxTokens { get; set; }
 
     [JsonProperty("temperature")]
-    public int? Temperature { get; set; }
+    public double? Temperature { get; set; }
 
     [JsonProperty("top_p")]
-    public int? TopP { get; set; }
+    public double? TopP { get; set; }
 
     [JsonProperty("logit_bias")]
     public string? LogitBias { get; set; }
@@ -47,8 +47,8 @@ public class CompletionRequest : Request
     public string? Stop { get; set; }
 
     [JsonProperty("presence_penalty")]
-    public int? PresencePenalty { get; set; }
+    public double? PresencePenalty { get; set; }
 
     [JsonProperty("frequency_penalty")]
-    public int? FrequencyPenalty { get; set; }
+    public double? FrequencyPenalty { get; set; }
 }
